(refactor) extract classes from MapsController
No unit tests yet, but it feels better to have these a bit more
organised
