+++ lib/day_names.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 module DayNames
+  class NonDayError < StandardError; end
+
   def self.name(date)
     I18n.l(date, format: '%A')
   end
